change msg query to sharding-jdbc;

+++ src/main/java/com/kuaizhan/kzweixin/controller/vo/MsgVO.java
@@ -18,5 +18,5 @@ public class MsgVO {
     private String headImgUrl;
     private String openId;
     private Map content;
-    private Long createTime;
+    private Integer createTime;
 }
